+++ app/Http/Filters/CocktailQueryFilter.php
@@ -23,14 +23,11 @@ final class CocktailQueryFilter extends QueryBuilder
 
         $barMembership = $this->request->user()->getBarMembership(bar()->id);
 
-        $useParentIngredientAsSubstitute = $barMembership->use_parent_as_substitute;
-
         $this
             ->allowedFilters([
                 AllowedFilter::exact('id'),
                 AllowedFilter::custom('name', new FilterNameSearch()),
                 AllowedFilter::partial('ingredient_name', 'ingredients.ingredient.name'),
-                // AllowedFilter::exact('ingredient_id', 'ingredients.ingredient.id'),
                 AllowedFilter::exact('ingredient_substitute_id', 'ingredients.substitutes.ingredient.id'),
                 AllowedFilter::callback('ingredient_id', function ($query, $value) {
                     if (!is_array($value)) {
@@ -63,17 +60,17 @@ final class CocktailQueryFilter extends QueryBuilder
                         $query->userFavorites($barMembership->id);
                     }
                 }),
-                AllowedFilter::callback('on_shelf', function ($query, $value) use ($useParentIngredientAsSubstitute) {
+                AllowedFilter::callback('on_shelf', function ($query, $value) {
                     if ($value === true) {
-                        $query->whereIn('cocktails.id', $this->request->user()->getShelfCocktailsOnce(bar()->id, $useParentIngredientAsSubstitute));
+                        $query->whereIn('cocktails.id', $this->request->user()->getShelfCocktailsOnce(bar()->id));
                     }
                 }),
-                AllowedFilter::callback('bar_shelf', function ($query, $value) use ($useParentIngredientAsSubstitute) {
+                AllowedFilter::callback('bar_shelf', function ($query, $value) {
                     if ($value === true) {
-                        $query->whereIn('cocktails.id', bar()->getShelfCocktailsOnce($useParentIngredientAsSubstitute));
+                        $query->whereIn('cocktails.id', bar()->getShelfCocktailsOnce());
                     }
                 }),
-                AllowedFilter::callback('user_shelves', function ($query, $value) use ($cocktailRepo, $useParentIngredientAsSubstitute) {
+                AllowedFilter::callback('user_shelves', function ($query, $value) use ($cocktailRepo, $barMembership) {
                     if (!is_array($value)) {
                         $value = [$value];
                     }
@@ -88,17 +85,17 @@ final class CocktailQueryFilter extends QueryBuilder
 
                     $query->whereIn('cocktails.id', $cocktailRepo->getCocktailsByIngredients(
                         $ingredients->pluck('ingredient_id')->toArray(),
-                        useParentIngredientAsSubstitute: $useParentIngredientAsSubstitute,
+                        $barMembership->bar_id,
                     ));
                 }),
-                AllowedFilter::callback('shelf_ingredients', function ($query, $value) use ($cocktailRepo, $useParentIngredientAsSubstitute) {
+                AllowedFilter::callback('shelf_ingredients', function ($query, $value) use ($cocktailRepo, $barMembership) {
                     if (!is_array($value)) {
                         $value = [$value];
                     }
 
                     $query->whereIn('cocktails.id', $cocktailRepo->getCocktailsByIngredients(
                         $value,
-                        useParentIngredientAsSubstitute: $useParentIngredientAsSubstitute
+                        $barMembership->bar_id,
                     ));
                 }),
                 AllowedFilter::callback('is_public', function ($query, $value) {
@@ -224,7 +221,7 @@ final class CocktailQueryFilter extends QueryBuilder
             })
             ->groupBy('cocktails.id')
             ->filterByBar('cocktails')
-            ->with('bar.shelfIngredients')
+            ->with(['bar.shelfIngredients', 'ingredients.ingredient.bar'])
             ->withRatings($this->request->user()->id);
     }
 }