+++ app/Http/Middleware/Authenticate.php
@@ -2,20 +2,17 @@
 
 namespace Kami\Cocktail\Http\Middleware;
 
+use Illuminate\Http\Request;
 use Illuminate\Auth\Middleware\Authenticate as Middleware;
 
 class Authenticate extends Middleware
 {
-    /**
-     * Get the path the user should be redirected to when they are not authenticated.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return string|null
-     */
-    protected function redirectTo($request)
+    protected function redirectTo(Request $request): ?string
     {
         if (!$request->expectsJson()) {
             return '/';
         }
+
+        return null;
     }
 }