+++ app/Http/Resources/CocktailIngredientResource.php
@@ -27,14 +27,42 @@ class CocktailIngredientResource extends JsonResource
             'optional' => (bool) $this->optional,
             'ingredient' => new IngredientBasicResource($this->ingredient),
             'substitutes' => CocktailIngredientSubstituteResource::collection($this->whenLoaded('substitutes')),
+            'variants_in_shelf' => $this->when($this->ingredient->relationLoaded('descendants'), fn () => IngredientBasicResource::collection($this->ingredient->barShelfVariants())),
             'note' => $this->note,
+            'is_specified' => (bool) $this->is_specified,
             'formatted' => new AmountFormats($this->resource),
-            'in_shelf' => $this->when($this->relationLoaded('substitutes'), fn () => $this->userHasInShelf($request->user())),
-            'in_shelf_as_substitute' => $this->when($this->relationLoaded('substitutes'), fn () => $this->userHasInShelfAsSubstitute($request->user())),
-            'in_shelf_as_complex_ingredient' => $this->when($this->relationLoaded('substitutes'), fn () => $this->userHasInShelfAsComplexIngredient($request->user())),
-            'in_bar_shelf' => $this->when($this->relationLoaded('substitutes'), fn () => $this->barHasInShelf()),
-            'in_bar_shelf_as_substitute' => $this->when($this->relationLoaded('substitutes'), fn () => $this->barHasInShelfAsSubstitute()),
-            'in_bar_shelf_as_complex_ingredient' => $this->when($this->relationLoaded('substitutes'), fn () => $this->barHasInShelfAsComplexIngredient()),
+            'in_shelf' => $this->when(
+                $this->relationLoaded('ingredient'),
+                fn () => $this->ingredient->userHasInShelf($request->user())
+            ),
+            'in_shelf_as_variant' => $this->when(
+                $this->ingredient->relationLoaded('descendants'),
+                fn () => !$this->is_specified && $this->ingredient->userShelfVariants($request->user())->count() > 0
+            ),
+            'in_shelf_as_substitute' => $this->when(
+                $this->relationLoaded('substitutes'),
+                fn () => $this->userHasInShelfAsSubstitute($request->user())
+            ),
+            'in_shelf_as_complex_ingredient' => $this->when(
+                $this->ingredient->relationLoaded('ingredientParts'),
+                fn () => $this->ingredient->userHasInShelfAsComplexIngredient($request->user())
+            ),
+            'in_bar_shelf' => $this->when(
+                $this->relationLoaded('ingredient'),
+                fn () => $this->ingredient->barHasInShelf()
+            ),
+            'in_bar_shelf_as_substitute' => $this->when(
+                $this->relationLoaded('substitutes'),
+                fn () => $this->barHasInShelfAsSubstitute()
+            ),
+            'in_bar_shelf_as_complex_ingredient' => $this->when(
+                $this->ingredient->relationLoaded('ingredientParts'),
+                fn () => $this->ingredient->barHasInShelfAsComplexIngredient()
+            ),
+            'in_bar_shelf_as_variant' => $this->when(
+                $this->ingredient->relationLoaded('descendants'),
+                fn () => !$this->is_specified && $this->ingredient->barShelfVariants()->count() > 0
+            ),
         ];
     }
 }