+++ database/migrations/2025_02_17_201733_remove_ingredient_categories.php
@@ -1,0 +1,37 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+return new class () extends Migration {
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::table('ingredients', function (Blueprint $table) {
+            $table->dropForeign(['ingredient_category_id']);
+            $table->dropColumn('ingredient_category_id');
+        });
+        Schema::dropIfExists('ingredient_categories');
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::create('ingredient_categories', function (Blueprint $table) {
+            $table->id();
+            $table->foreignId('bar_id')->constrained()->onDelete('cascade');
+            $table->string('name');
+            $table->text('description')->nullable();
+            $table->timestamps();
+        });
+
+        Schema::table('ingredients', function (Blueprint $table) {
+            $table->foreignId('ingredient_category_id')->nullable()->constrained()->nullOnDelete();
+        });
+    }
+};