+++ tests/Feature/Http/ImportControllerTest.php
@@ -6,7 +6,6 @@ namespace Tests\Feature\Http;
 
 use Tests\TestCase;
 use Illuminate\Http\UploadedFile;
-use Kami\Cocktail\Models\IngredientCategory;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 
 class ImportControllerTest extends TestCase
@@ -50,10 +49,6 @@ class ImportControllerTest extends TestCase
 
         $this->withHeader('Bar-Assistant-Bar-Id', (string) $membership->bar_id);
 
-        $existingCategory = IngredientCategory::factory()->for($membership->bar)->create([
-            'name' => 'Liquers'
-        ]);
-
         $source = file_get_contents(base_path('tests/fixtures/ingredients.csv'));
 
         $response = $this->postJson('/api/import/ingredients', [
@@ -63,11 +58,9 @@ class ImportControllerTest extends TestCase
         $response->assertSuccessful();
 
         $this->assertDatabaseCount('ingredients', 4);
-        $this->assertDatabaseCount('ingredient_categories', 2);
         $this->assertDatabaseHas('ingredients', [
             'name' => 'Campari',
             'slug' => 'campari-' . $membership->bar_id,
-            'ingredient_category_id' => $existingCategory->id,
             'strength' => 40,
             'description' => 'Bitter liquer',
             'origin' => 'Italy',
@@ -78,7 +71,6 @@ class ImportControllerTest extends TestCase
         $this->assertDatabaseHas('ingredients', [
             'name' => 'gin',
             'slug' => 'gin-' . $membership->bar_id,
-            'ingredient_category_id' => 2,
             'strength' => 23.3,
             'description' => null,
             'origin' => null,
@@ -89,7 +81,6 @@ class ImportControllerTest extends TestCase
         $this->assertDatabaseHas('ingredients', [
             'name' => 'Whiskey',
             'slug' => 'whiskey-' . $membership->bar_id,
-            'ingredient_category_id' => 2,
             'strength' => 0,
             'description' => null,
             'origin' => null,
@@ -100,7 +91,6 @@ class ImportControllerTest extends TestCase
         $this->assertDatabaseHas('ingredients', [
             'name' => 'empty',
             'slug' => 'empty-' . $membership->bar_id,
-            'ingredient_category_id' => null,
             'strength' => 0,
             'description' => null,
             'origin' => null,